everything works while method handleTransaction doesn't

+++ src/main/java/com/walletapp/task/entities/Wallet.java
@@ -18,5 +18,5 @@ public class Wallet {
 
     @Getter
     @Setter
-    private Long balance;
+    private Float balance;
 }
